Rename subfolder for Lists
